[vulkan] Split vulkan/command file.
This CL splits the vulkan/command.{h|cc} into command_buffer and
command_pool files.
