+++ src/vulkan/command_buffer.cc
@@ -1,4 +1,4 @@
-// Copyright 2018 The Amber Authors.
+// Copyright 2019 The Amber Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,41 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "src/vulkan/command.h"
-
-#include <memory>
-
-#include "src/make_unique.h"
-#include "src/vulkan/device.h"
+#include "src/vulkan/command_buffer.h"
 
 namespace amber {
-
 namespace vulkan {
-
-CommandPool::CommandPool(Device* device) : device_(device) {}
-
-CommandPool::~CommandPool() = default;
-
-Result CommandPool::Initialize(uint32_t queue_family_index) {
-  VkCommandPoolCreateInfo pool_info = VkCommandPoolCreateInfo();
-  pool_info.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
-  pool_info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-  pool_info.queueFamilyIndex = queue_family_index;
-
-  if (device_->GetPtrs()->vkCreateCommandPool(device_->GetDevice(), &pool_info,
-                                              nullptr, &pool_) != VK_SUCCESS) {
-    return Result("Vulkan::Calling vkCreateCommandPool Fail");
-  }
-
-  return {};
-}
-
-void CommandPool::Shutdown() {
-  if (pool_ != VK_NULL_HANDLE) {
-    device_->GetPtrs()->vkDestroyCommandPool(device_->GetDevice(), pool_,
-                                             nullptr);
-  }
-}
 
 CommandBuffer::CommandBuffer(Device* device, VkCommandPool pool, VkQueue queue)
     : device_(device), pool_(pool), queue_(queue) {}
@@ -149,5 +118,4 @@ void CommandBuffer::Shutdown() {
 }
 
 }  // namespace vulkan
-
 }  // namespace amber
