+++ src/vulkan/engine_vulkan.h
@@ -23,7 +23,7 @@
 #include "amber/vulkan_header.h"
 #include "src/cast_hash.h"
 #include "src/engine.h"
-#include "src/vulkan/command.h"
+#include "src/vulkan/command_pool.h"
 #include "src/vulkan/device.h"
 #include "src/vulkan/pipeline.h"
 #include "src/vulkan/vertex_buffer.h"
