+++ src/vulkan/pipeline.h
@@ -24,7 +24,7 @@
 #include "amber/vulkan_header.h"
 #include "src/cast_hash.h"
 #include "src/engine.h"
-#include "src/vulkan/command.h"
+#include "src/vulkan/command_buffer.h"
 #include "src/vulkan/descriptor.h"
 #include "src/vulkan/push_constant.h"
